Servo dreht in 10° von 0-90° Schritten und bleibt am Schluss beim nächsten erkannten Töggel

+++ SW/Peco/src/WandDistanz/WandDistanz.h
@@ -8,7 +8,7 @@ class WandDistanz
   public:
 	WandDistanz();
     void init(Print &print, int p_iDistanceSensorInputPIN);
-	int WandDistanz::getAktuelleDistanzCm();
+	long WandDistanz::getAktuelleDistanzCm();
 
   private:
   	Print* printer;
